+++ src/main/java/software/works/fizzbuzz/FizzBuzzWordsToList.java
@@ -1,29 +1,29 @@
 package software.works.fizzbuzz;
 
 import java.util.List;
-import java.util.stream.IntStream;
+import java.util.stream.LongStream;
 
 import software.works.fizzbuzz.rule.PlayerBuilder;
 
 class FizzBuzzWordsToList {
 
     private PlayerBuilder playerBuilder;
-    private int start;
+    private long start;
 
     FizzBuzzWordsToList(PlayerBuilder playerBuilder) {
         this.playerBuilder = playerBuilder;
     }
 
-    public List<String> of(int... values) {
+    public List<String> of(long... values) {
         return playerBuilder.chosenPlayer().playAtFizzBuzzToList(values);
     }
 
-    public FizzBuzzWordsToList from(int start) {
+    public FizzBuzzWordsToList from(long start) {
         this.start = start;
         return this;
     }
 
     public List<String> to(int end) {
-        return playerBuilder.chosenPlayer().playAtFizzBuzzToList(IntStream.rangeClosed(start, end).toArray());
+        return playerBuilder.chosenPlayer().playAtFizzBuzzToList(LongStream.rangeClosed(start, end).toArray());
     }
 }
